requirements.txt
speedtest-cli
requests
dig
